\name{fd_breaks}
\alias{fd_breaks}
\title{Pick breaks automatically using the Freedman-Diaconis rule...}

\description{
  Pick breaks automatically using the Freedman-Diaconis
  rule
}
\value{A function that takes a single parameter, a numeric x specifying
the data for which breaks are needed, and returns a vector of breaks.}
