\name{scott_breaks}
\alias{scott_breaks}
\title{Pick breaks automatically using Scott's rule...}

\description{
  Pick breaks automatically using Scott's rule
}
\value{A function that takes a single parameter, a numeric x specifying
the data for which breaks are needed, and returns a vector of breaks.}
