\name{sturges_breaks}
\alias{sturges_breaks}
\title{Pick breaks automatically using Sturges' rule...}

\description{
  Pick breaks automatically using Sturges' rule
}
\value{A function that takes a single parameter, a numeric x specifying
the data for which breaks are needed, and returns a vector of breaks.}
